backend/requirements.txt
asgiref==3.7.1
click==8.1.3
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-flex-fields==1.0.2
gunicorn==20.1.0
h11==0.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
uvicorn==0.22.0
